requirements.txt
PySide6>=6.6.0
pydantic>=2.0.0
PyYAML>=6.0
requests>=2.31.0
soundfile>=0.12.0
numpy>=1.24.0
opencc>=1.2.0
